Move all UI config groups into a subdirectory.
