+++ src/invoke_training/ui/config_groups/optimizer_config_group.py
@@ -3,7 +3,7 @@ from typing import Any
 import gradio as gr
 
 from invoke_training.config.optimizer.optimizer_config import AdamOptimizerConfig, ProdigyOptimizerConfig
-from invoke_training.ui.ui_config_element import UIConfigElement
+from invoke_training.ui.config_groups.ui_config_element import UIConfigElement
 
 OptimizerConfig = AdamOptimizerConfig | ProdigyOptimizerConfig
